+++ app/src/main/java/com.terracode.blueharvest/listeners/operationValueListeners/HeightCoefficientListener.kt
@@ -23,7 +23,7 @@ class HeightCoefficientListener(
 
     //Constants
     private val configName = ContextCompat.getString(activity, R.string.coefficientHeightTitle)
-    private var maxUserInput = MaxUserInputInt.MAX_HEIGHT_INPUT.value.toDouble()
+    private var maxUserInput = MaxUserInputInt.MAX_HEIGHT_INPUT.value.toFloat()
 
     //Current Value
     private val unitToggle = PreferenceManager.getSelectedUnit()
