+++ app/src/main/java/com.terracode.blueharvest/listeners/optimalRangeValueListeners/OptimalHeightRangeListener.kt
@@ -24,7 +24,7 @@ class OptimalHeightRangeListener(
 
     //Constants
     private val configName = ContextCompat.getString(activity, R.string.optimalHeightRangeTitle)
-    private var maxUserInput = MaxUserInputInt.MAX_HEIGHT_INPUT.value.toDouble()
+    private var maxUserInput = MaxUserInputInt.MAX_HEIGHT_INPUT.value.toFloat()
 
     //Current Value
     private val unitToggle = PreferenceManager.getSelectedUnit()
